+++ .env.dev.local.symfony.example
@@ -1,0 +1,12 @@
+# .env.dev.local
+APP_ENV=dev
+
+APP_DOMAIN=dev.corbisier.fr
+
+AUTH_HOST=auth.dev.corbisier.fr
+API_HOST=api.dev.corbisier.fr
+
+AUTH_URL=https://auth.dev.corbisier.fr
+API_URL=https://api.dev.corbisier.fr
+
+CORS_ALLOW_ORIGIN=^https://(auth\.dev\.corbisier\.fr|api\.dev\.corbisier\.fr|corbycats\.fr)$
