+++ .env.docker-compose.example
@@ -33,11 +33,22 @@ MYSQL_DATABASE=${NAME}
 MYSQL_USER=user
 MYSQL_PASSWORD=password
 
+MYSQL_HOST_API=database_api
+MYSQL_URL_API=127.0.0.1
+MYSQL_ROOT_PASSWORD_API=P@ssW0rd!
+MYSQL_DATABASE_API=${NAME}_API
+MYSQL_USER_API=user
+MYSQL_PASSWORD_API=password
+
+
 ###> PORTS ###
 # Ajustez les ports LOCALHOST si nécessaire pour éviter
 # les conflits avec d'autres services sur votre machine
 SQL_DOCKER_PORT=3306
 SQL_LOCALHOST_PORT=3376
+
+SQL_DOCKER_PORT_API=3306
+SQL_LOCALHOST_PORT_API=3366
 
 PHPMYADMIN_LOCALHOST_PORT=8081
 PHPMYADMIN_DOCKER_PORT=80
@@ -48,8 +59,12 @@ ADMINER_LOCALHOST_PORT=8070
 
 ###> FOLDERS ###
 FOLDER_DATA=database
+
 FOLDER_DATASQL=${FOLDER_DATA}/sql
 FOLDER_DATABASE=${FOLDER_DATA}/${DATABASE_NAME}
+
+FOLDER_DATASQL_API=${FOLDER_DATA}/sql_api
+FOLDER_DATABASE_API=${FOLDER_DATA}/${DATABASE_NAME}_api
 ###< FOLDERS ###
