+++ .env.local.symfony.example
@@ -1,0 +1,13 @@
+# .env.local
+
+APP_ENV=dev
+
+APP_DOMAIN=localhost
+
+AUTH_HOST=localhost
+API_HOST=localhost
+
+AUTH_URL=http://localhost:8000
+API_URL=http://localhost:8000
+
+CORS_ALLOW_ORIGIN=^http://localhost(:[0-9]+)?$
