+++ .env.prod.local.symfony.example
@@ -1,0 +1,12 @@
+# .env.prod.local
+APP_ENV=prod
+
+APP_DOMAIN=corbisier.fr
+
+AUTH_HOST=auth.corbisier.fr
+API_HOST=api.corbisier.fr
+
+AUTH_URL=https://auth.corbisier.fr
+API_URL=https://api.corbisier.fr
+
+CORS_ALLOW_ORIGIN=^https://(auth\.corbisier\.fr|api\.corbisier\.fr|corbycats\.fr)$
