+++ api/.env.symfony.example
@@ -1,0 +1,96 @@
+# In all environments, the following files are loaded if they exist,
+# the latter taking precedence over the former:
+#
+#  * .env                contains default values for the environment variables needed by the app
+#  * .env.local          uncommitted file with local overrides
+#  * .env.$APP_ENV       committed environment-specific defaults
+#  * .env.$APP_ENV.local uncommitted environment-specific overrides
+#
+# Real environment variables win over .env files.
+#
+# DO NOT DEFINE PRODUCTION SECRETS IN THIS FILE NOR IN ANY OTHER COMMITTED FILES.
+# https://symfony.com/doc/current/configuration/secrets.html
+#
+# Run "composer dump-env prod" to compile .env files for production use (requires symfony/flex >=1.2).
+# https://symfony.com/doc/current/best_practices.html#use-environment-variables-for-infrastructure-configuration
+####> PROJECT INFO ###
+NAME=symfony-api
+BASE=mariadb
+###< PROJECT INFO ###
+
+###> VERSIONS ###
+BACKEND_VERSION=1.0.0
+MARIADB_VERSION=11.7.0
+###< VERSIONS ###
+
+###> MYSQL/ mariadb - adminer ###
+DATABASE_NAME=${NAME}_${BASE}
+MYSQL_HOST=database
+MYSQL_URL=127.0.0.1
+MYSQL_ROOT_PASSWORD=P@ssW0rd!
+MYSQL_DATABASE=${NAME}
+MYSQL_USER=user
+MYSQL_PASSWORD=password
+###< MYSQL/ mariadb - adminer ###
+
+###> PORTS ###
+# Ajustez les ports LOCALHOST si nécessaire pour éviter
+# les conflits avec d'autres services sur votre machine
+SQL_DOCKER_PORT=3306
+SQL_LOCALHOST_PORT=3376
+###< PORTS ###
+
+
+###> symfony/framework-bundle ###
+APP_ENV=dev
+APP_SECRET=<VOTRE_APP_SECRET>
+APP_SHARE_DIR=var/share
+###< symfony/framework-bundle ###
+
+###> symfony/routing ###
+# Configure how to generate URLs in non-HTTP contexts, such as CLI commands.
+# See https://symfony.com/doc/current/routing.html#generating-urls-in-commands
+DEFAULT_URI=http://api.corbisier.test
+###< symfony/routing ###
+
+###> doctrine/doctrine-bundle ###
+# Format described at https://www.doctrine-project.org/projects/doctrine-dbal/en/latest/reference/configuration.html#connecting-using-a-url
+# IMPORTANT: You MUST configure your server version, either here or in config/packages/doctrine.yaml
+#
+# DATABASE_URL="sqlite:///%kernel.project_dir%/var/data_%kernel.environment%.db"
+# DATABASE_URL="mysql://app:!ChangeMe!@127.0.0.1:3306/app?serverVersion=8.0.32&charset=utf8mb4"
+DATABASE_URL="mysql://${MYSQL_USER}:${MYSQL_PASSWORD}@${MYSQL_URL}:${SQL_LOCALHOST_PORT}/${MYSQL_DATABASE}?serverVersion=${MARIADB_VERSION}-MariaDB&charset=utf8mb4"
+# DATABASE_URL="postgresql://app:!ChangeMe!@127.0.0.1:5432/app?serverVersion=16&charset=utf8"
+###< doctrine/doctrine-bundle ###
+
+###> nelmio/cors-bundle ###
+# SI UN SEUL SITE
+# CORS_ALLOW_ORIGIN="^${DEFAULT_URI}|(https?://(localhost|127\.0\.0\.1)(:[0-9]+)?)$"
+
+# LOCAL
+CORS_ALLOW_ORIGIN=^http://localhost(:[0-9]+)?$
+
+# SUR SERVEUR DEVELOP
+# CORS_ALLOW_ORIGIN=^https?://(dev-.*\.corbisier\.fr|localhost(:[0-9]+)?)$
+
+# SUR PROD
+# CORS_ALLOW_ORIGIN=^https://(auth\.corbisier\.fr|api\.corbisier\.fr|corbycats\.fr)$
+
+
+CORS_ALLOW_METHODS="GET,POST,PUT,DELETE,OPTIONS"
+CORS_ALLOW_HEADERS="Content-Type,Authorization"
+CORS_EXPOSE_HEADERS="Link"
+CORS_MAX_AGE=3600
+CORS_ALLOW_CREDENTIALS=true
+###< nelmio/cors-bundle ###
+
+
+###> lexik/jwt-authentication-bundle ###
+JWT_SECRET_KEY=%kernel.project_dir%/config/jwt/private.pem
+JWT_PUBLIC_KEY=%kernel.project_dir%/config/jwt/public.pem
+JWT_PASSPHRASE=<VOTRE_PASSPHRASE>
+###< lexik/jwt-authentication-bundle ###
+
+###> TENANT CONFIG ###
+TENANT_ALLOWED_DOMAINS="https://ton-site.fr,https://api.autre-site.fr"
+###< TENANT CONFIG ###
