+++ api/config/packages/security.yaml
@@ -1,39 +1,56 @@
 security:
-    # https://symfony.com/doc/current/security.html#registering-the-user-hashing-passwords
-    password_hashers:
-        Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface: 'auto'
 
-    # https://symfony.com/doc/current/security.html#loading-the-user-the-user-provider
+    password_hashers:
+        App\Auth\Entity\User:
+            algorithm: auto
+
     providers:
-        users_in_memory: { memory: null }
+        auth_user_provider:
+            entity:
+                class: App\Auth\Entity\User
+                property: email
 
     firewalls:
+
+        # 🔧 DEV (profiler, assets…)
         dev:
-            # Ensure dev tools and static assets are always allowed
             pattern: ^/(_profiler|_wdt|assets|build)/
             security: false
-        main:
+
+        # 🔐 API (stateless JWT)
+        api:
+            pattern: ^/api
+            stateless: true
+            jwt: ~
+
+        # 🔓 AUTH (login + session)
+        auth:
+            pattern: ^/
+            provider: auth_user_provider
             lazy: true
-            provider: users_in_memory
+            stateless: false
 
-            # Activate different ways to authenticate:
-            # https://symfony.com/doc/current/security.html#the-firewall
+            form_login:
+                login_path: /login
+                check_path: /login
 
-            # https://symfony.com/doc/current/security/impersonating_user.html
-            # switch_user: true
+            logout:
+                path: /logout
 
-    # Note: Only the *first* matching rule is applied
     access_control:
-        # - { path: ^/admin, roles: ROLE_ADMIN }
-        # - { path: ^/profile, roles: ROLE_USER }
 
-when@test:
-    security:
-        password_hashers:
-            # Password hashers are resource-intensive by design to ensure security.
-            # In tests, it's safe to reduce their cost to improve performance.
-            Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface:
-                algorithm: auto
-                cost: 4 # Lowest possible value for bcrypt
-                time_cost: 3 # Lowest possible value for argon
-                memory_cost: 10 # Lowest possible value for argon
+        # =========================
+        # 🔐 API (PRIORITAIRE)
+        # =========================
+        - { path: ^/api/logs, roles: ROLE_LOGS }
+        # - { path: ^/api/corbycats, roles: ROLE_CAT }
+
+        # =========================
+        # 🔓 AUTH
+        # =========================
+        - { path: ^/login, roles: PUBLIC_ACCESS }
+
+        # =========================
+        # 🔒 AUTRES ROUTES AUTH
+        # =========================
+        - { path: ^/, roles: IS_AUTHENTICATED_FULLY }
