+++ api/config/services.yaml
@@ -1,26 +1,32 @@
-# yaml-language-server: $schema=../vendor/symfony/dependency-injection/Loader/schema/services.schema.json
+# config/services.yaml
 
-# This file is the entry point to configure your own services.
-# Files in the packages/ subdirectory configure your dependencies.
-# See also https://symfony.com/doc/current/service_container/import.html
-
-# Put parameters here that don't need to change on each machine where the app is deployed
-# https://symfony.com/doc/current/best_practices.html#use-parameters-for-application-configuration
 parameters:
 
 services:
-    # default configuration for services in *this* file
-    _defaults:
-        autowire: true      # Automatically injects dependencies in your services.
-        autoconfigure: true # Automatically registers your services as commands, event subscribers, etc.
 
-    # makes classes in src/ available to be used as services
-    # this creates a service per class whose id is the fully-qualified class name
+    _defaults:
+        autowire: true
+        autoconfigure: true
+
+    # 🔥 Chargement global
     App\:
         resource: '../src/'
-    App\Logs\Context\TenantProvider:
+        exclude:
+            - '../src/DependencyInjection/'
+            - '../src/Kernel.php'
+
+    # =========================
+    # 🧠 Cache Tenant
+    # =========================
+    tenant.cache:
+        class: Symfony\Component\Cache\Adapter\FilesystemAdapter
+        arguments:
+            $namespace: 'tenant'
+            $defaultLifetime: 3600
+
+    # =========================
+    # 📊 Logs - TenantProvider
+    # =========================
+    App\Api\Logs\Context\TenantProvider:
         arguments:
             $cache: '@tenant.cache'
-
-    # add more service definitions when explicit configuration is needed
-    # please note that last definitions always *replace* previous ones
