+++ api/src/Api/Logs/Controller/TestController.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Api\Logs\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\Routing\Annotation\Route;
+
+class TestController extends AbstractController
+{
+    #[Route('/test', name: 'api_logs_test', methods: ['GET'])]
+    public function test(): JsonResponse
+    {
+        return $this->json([
+            'status' => 'ok',
+            'service' => 'api_logs',
+            'message' => 'API Logs route works',
+        ]);
+    }
+}
