+++ api/src/Api/Logs/Entity/LogDomain.php
@@ -1,28 +1,31 @@
 <?php
-namespace App\Logs\Entity;
+namespace App\Api\Logs\Entity;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 
 #[ORM\Entity]
-#[ORM\Table(name: 'CBV_LOGS_URI')]
+#[ORM\Table(name: 'CBV_LOGS_DOMAIN')]
 /**
- * URI normalisée ciblée par un evenement de log.
+ * Domaine fonctionnel rattache a une origine de log.
  */
-class LogUri {
+class LogDomain {
     /**
-     * Identifiant technique de l'URI.
+     * Identifiant technique du domaine.
      */
     #[ORM\Id, ORM\GeneratedValue, ORM\Column(type: 'bigint')]
     private ?int $id = null;
 
     /**
-     * Chemin unique de la ressource appelee.
+     * URL canonique utilisee pour le scoping tenant.
      */
     #[ORM\Column(length: 255, unique: true)]
-    private string $uri;
+    private string $url;
 
-    #[ORM\OneToMany(mappedBy: 'uri', targetEntity: LogOrigin::class)]
+    #[ORM\Column(type: 'boolean', options: ['default' => true])]
+    private bool $isActive = true;
+
+    #[ORM\OneToMany(mappedBy: 'domain', targetEntity: LogOrigin::class)]
     private Collection $origins;
 
     public function __construct()
@@ -35,14 +38,26 @@ class LogUri {
         return $this->id;
     }
 
-    public function getUri(): string
+    public function getUrl(): string
     {
-        return $this->uri;
+        return $this->url;
     }
 
-    public function setUri(string $uri): self
+    public function setUrl(string $url): self
     {
-        $this->uri = trim($uri);
+        $this->url = strtolower(rtrim(trim($url), '/'));
+
+        return $this;
+    }
+
+    public function isActive(): bool
+    {
+        return $this->isActive;
+    }
+
+    public function setIsActive(bool $isActive): self
+    {
+        $this->isActive = $isActive;
 
         return $this;
     }
@@ -73,6 +88,6 @@ class LogUri {
 
     public function __toString(): string
     {
-        return isset($this->uri) ? $this->uri : '';
+        return isset($this->url) ? $this->url : '';
     }
 }
